Fixes #5. Don't start a socket server during integration tests; use supertest Express support
