+++ package.json
@@ -54,7 +54,7 @@
   },
   "scripts": {
     "start": "node .\\server.js",
-    "test": "mocha -R xunit-file",
+    "test": "mocha --recursive test --reporter-options=xunit.xml",
     "ci-full": "istanbul cover _mocha -- -R xunit-file",
     "test-windows": "./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha/ -- -R xunit-file && istanbul report html"
   }
